Release tree in StageSelectScene without viewing LevelSelectScene
